+++ packages/form/src/index.tsx
@@ -19,7 +19,6 @@ export type {
   LightWrapperProps,
   ProFormInstance,
 } from './BaseForm';
-export { FieldContext } from './FieldContext';
 export * from './components';
 export type {
   FormListActionType,
@@ -44,6 +43,7 @@ export type {
   ProFormUploadDraggerProps,
 } from './components';
 export { FormListContext } from './components/List';
+export { FieldContext } from './FieldContext';
 export * from './layouts';
 export type {
   LightFilterFooterRender,